quizAi/requirements.txt
requests>=2.31.0
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python
google-cloud-aiplatform
